\name{available_phage_palettes}
\alias{available_phage_palettes}
\title{List available ColorPhage palettes}
\usage{
available_phage_palettes()
}
\value{
A data frame with palette metadata.
}
\description{
Lists the currently available ColorPhage simple-mode palettes.
}
\examples{
available_phage_palettes()
}
